requirements.txt
numpy==1.25.2
opencv-python==4.8.0.76
pyvirtualcam=0.10.2
Pillow==10.0.0